Update usage of setup_bin
The attribute "setup_bin" was not used up to the latest cookbook version:
the mvn installation directory was systematically added in the environment
PATH.

This commit adds the option to symlink mvn in the usual binary directory
through an ark resource. It also displaces "setup_bin" from its previous
stated usage to configuring this option.